fixes bug creating a dependency builder each time
