+++ src/main/java/org/jboss/forge/arquillian/command/algeron/AddAlgeronUrlPublisher.java
@@ -1,5 +1,7 @@
-package org.jboss.forge.arquillian.command;
+package org.jboss.forge.arquillian.command.algeron;
 
+import org.jboss.forge.addon.facets.FacetFactory;
+import org.jboss.forge.addon.projects.ProjectFactory;
 import org.jboss.forge.addon.ui.context.UIBuilder;
 import org.jboss.forge.addon.ui.context.UIContext;
 import org.jboss.forge.addon.ui.input.UIInput;
@@ -12,12 +14,18 @@ import javax.inject.Inject;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
-public class AddAlgeronFolderPublisher extends AbstractAlgeronPublisherCommand
+public class AddAlgeronUrlPublisher extends AbstractAlgeronPublisherCommand
 {
 
    @Inject
-   @WithAttributes(shortName = 'o', label = "Output Folder", required = true)
-   private UIInput<String> outputFolder;
+   private ProjectFactory projectFactory;
+
+   @Inject
+   private FacetFactory facetFactory;
+
+   @Inject
+   @WithAttributes(shortName = 'u', label = "Url to POST", required = true)
+   private UIInput<String> url;
 
    @Inject
    @WithAttributes(shortName = 'c', label = "Contracts Folder")
@@ -28,14 +36,14 @@ public class AddAlgeronFolderPublisher extends AbstractAlgeronPublisherCommand
    {
       return Metadata.from(super.getMetadata(context), getClass())
               .category(Categories.create("Algeron"))
-              .name("Arquillian Algeron: Add Folder Publisher")
-              .description("This command registers a Folder Publisher for Algeron");
+              .name("Arquillian Algeron: Add Url Publisher")
+              .description("This command registers a Url Publisher for Algeron");
    }
 
    @Override
    public void initializeUI(UIBuilder builder) throws Exception
    {
-      builder.add(outputFolder).add(contractFolder);
+      builder.add(url).add(contractFolder);
       contractFolder.setDefaultValue("target/pacts");
    }
 
@@ -43,14 +51,14 @@ public class AddAlgeronFolderPublisher extends AbstractAlgeronPublisherCommand
    protected Map<String, String> getParameters()
    {
       Map<String, String> parameters = new LinkedHashMap<>();
-      parameters.put("provider", "folder");
-      parameters.put("outputFolder", outputFolder.getValue());
+      parameters.put("provider", "url");
+      parameters.put("url", url.getValue());
       parameters.put("contractsFolder", contractFolder.getValue());
       return parameters;
    }
 
    @Override
    protected String getName() {
-      return "Folder";
+      return "Url";
    }
 }
