+++ src/test/java/test/integration/support/ShellTestTemplate.java
@@ -45,7 +45,6 @@ public abstract class ShellTestTemplate {
         }
         if (projectFactory != null) {
             projectFactory.invalidateCaches();
-            project.getRoot().delete(true);
         }
 
     }
